Api/requirements.txt
fastapi<0.100
uvicorn[standard]
python-dotenv
pydantic<2.0
lgpio
apscheduler==3.10.4
